[Bugfix] Rewrite peer message handling
Up until now messages were handled with every 'data' event of a peer.
This could fail on fragmented TCP messages.

Now we add all new peer messages to a buffer and check, if
the message was received in its entirety, then evaluate it.

Also remove redundant serializedSize entry in payloads because we have this
data in the message header.